# Various performance enhancements by making Julia code more idiomatic