… src/deterministic_vi/elbo_kl.jl
… src/deterministic_vi/elbo_objective.jl
… src/deterministic_vi/maximize_elbo.jl
… src/deterministic_vi/source_brightness.jl
… src/model/fsm_util.jl
… src/model/log_prob.jl
… src/model/wcs_utils.jl
… test/DerivativeTestUtils.jl
… test/test_derivatives.jl
… test/test_elbo.jl
… test/test_misc.jl
… test/test_optimization.jl
… test/test_psf.jl
… test/test_transforms.jl