This PR makes a bunch of small changes and a few bigger changes in order to make the Julia code more idiomatic, and thus increase performance. Here's a summary of the changes:

- refactor some allocating methods to reuse memory when possible
- refactor some methods to avoid redundant computations
- avoid using types like classes (e.g. remove `Function` fields)
- manually hoist pointer loads (Julia's compiler isn't smart enough yet to determine when a binding won't get de-aliased, so you have to manually do this if you don't want to duplicate loads)
- the `v` field of the `SensitiveFloat` type is now `::Base.RefValue{NumType}` instead of `::Vector{NumType}`. I initially tried just making it `NumType`, which would be the idiomatic thing to do, but that ended up making things very slow for some unknown reason (I believe @jeff-regier already knows about this).

I mainly touched `PSF.jl`, `Transform.jl`, and the files in `src/deterministic_vi`. The modifications to most of the other files are due to find-and-replace (e.g. `.v[1]` --> `.v[]`). There are other places in the code that could've benefited from the above changes as well, but this PR was already really large and I didn't want to go further overboard.

I ran `benchmark_infer.jl`, and it seems this PR gives a ~25% speedup with a single thread, and a ~30% speedup with 4 threads (running on Julia v0.6):

```
on master (1 thread): 4.432900 seconds (1.78 M allocations: 316.157 MB, 6.07% gc time)
on master (4 threads): 3.076258 seconds (1.69 M allocations: 312.785 MB, 16.95% gc time)

on jr/idiomatic2 (1 thread): 3.526573 seconds (1.95 M allocations: 284.098 MB, 7.39% gc time)
on jr/idiomatic2 (4 threads): 2.358118 seconds (1.80 M allocations: 279.312 MB, 11.32% gc time)
```

I haven't made much effort in getting tests to match the new code, so they probably won't pass until I fix them. At the very least, `benchmark_infer.jl` is giving me close to the same output on both this branch and master, so hopefully it'll just be a matter of code churn rather than bug fixing.