@jeff-regier 2016-11-14
Hi Jarrett, Great changes, thanks for the contribution. I ran `validate_on_stripe82.jl` to confirm that none of them change the scientific results. (They don't.) I ran `benchmark_rcf.jl` to get another perspective on how this will change runtime.

On master:
```
104.746012 seconds (95.70 M allocations: 8.689 GB, 2.39% gc time)
```

With this PR:
```
 90.919500 seconds (103.66 M allocations: 8.699 GB, 4.73% gc time)
```

The 15% decrease in run time is a big deal. That should save us around 50,000 cpu hours later this week when I do a large scale run.

Any idea why the number of allocations increase with this PR? Perhaps a lot of really small allocations somewhere? (The number of allocations grows by 10%, but the amount allocated grows by only 1%.)