@jrevels 2016-11-14
> Any idea why the number of allocations increase with this PR? Perhaps a lot of really small allocations somewhere? (The number of allocations grows by 10%, but the amount allocated grows by only 1%.)

Yeah, I noticed that as well. Before I pulled `master` and rebased, this PR reduced allocations vs. `master` from a couple days ago - now it seems the tides have turned. I probably slipped up somewhere, I'll see if I can pinpoint where the extra allocations are coming from.