@jeff-regier 2016-11-17
@jrevels  The main bug is in `deterministic_vi/source_brightness.jl`. I didn't find the particular line, but I swapped files in and out until I'd isolated it. (You just have to do a find and replace setting `v[1]` to `v[]`.)

`source_brightness.jl` is not cpu intensive at all, so no need to make prevent these pointer lookups. I'd propose substituting the old `source_brightness.jl` back in, with `v[1]` replaced with `v[]`.