@jrevels 2016-11-17
> source_brightness.jl is not cpu intensive at all, so no need to prevent pointer lookups there. I'd propose substituting the old source_brightness.jl back in, with v[1] replaced with v[].

Done! Should I wait until #418 is merged to master and then rebase, or are you planning to #418 be merged after this gets merged?