@jeff-regier 2016-11-17
I think PR #418 ended up being largely orthogonal to this one. The tests should pass on your pr without it, by rolling back `source_brightness.jl` and fixing the bugs in `Transform` I commented on. So let's merge this first. Will you figure out the extra memory allocations before merging this PR though?