@jrevels 2016-11-18
It turns out the extra allocations were due to an inference bug (ref JuliaLang/julia#15276). Fiddling around with type annotations fixes most of the additional allocations (but not all of them).