@jeff-regier 2016-11-18
With `benchmark_rcf.jl`, now I see
```
 92.319823 seconds (102.73 M allocations: 8.677 GB, 5.06% gc time)
```
So the total size of the allocations is now lower than before the PR, but the number of allocations is still higher. The number of allocations seems to matter more for gc time in this case. Thanks for continuing to work on this PR....the last 10% of the programming always seems to take longest....