@jrevels 2016-11-23
Awesome, thanks for handling that!

> Memory allocation on the merged branch is really low now, for reasons I don't fully understand.

I'm not too surprised - a lot of the changes here really do decrease local memory usage significantly (demonstrable via profiling with `--track-allocation`). The fact that these local improvements didn't present an impact in the overall allocation count seemed to be due to buggy type inference and small-but-frequent allocations that hit this branch particularly hard for some reason. It's reasonable to me that changes to master enabled the benefits of this branch to "propagate" more easily by improving upstream/downstream code. Of course, this is just a guess...and Julia's type inference can be a finicky beast to make assumptions about.